+++ NEWS.md
@@ -8,6 +8,7 @@
 
 - The default for the `mtp` argument has been changed from `FALSE` to `TRUE` (see issue \#170).
 - A warning message is now printed in `lmtp_contrast()` that p-values aren't adjusted for multiple comparisons (see issue \#172).
+- A warning message is now printed in `lmtp_contrast()` if `ref` is a constant.
 
 # lmtp 1.5.1